django/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
click==7.1.2
Django==3.1.3
djangorestframework==3.12.2
geographiclib==1.50
geopy==2.0.0
mypy-extensions==0.4.3
pathspec==0.8.1
python-dotenv==0.15.0
pytz==2020.4
regex==2020.11.13
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
